Rename Digesty -> Digest
