+++ demo/primitives/src/lib.rs
@@ -49,4 +49,4 @@ pub type Index = u64;
 pub type Hash = primitives::H256;
 
 /// Alias to 512-bit hash when used in the context of a signature on the relay chain.
-pub type Signature = runtime_primitives::generic::Ed25519Signature;
+pub type Signature = runtime_primitives::traits::Ed25519Signature;
