+++ substrate/runtime/council/src/lib.rs
@@ -35,7 +35,7 @@ extern crate substrate_runtime_staking as staking;
 extern crate substrate_runtime_system as system;
 
 use rstd::prelude::*;
-use primitives::{Zero, One, RefInto, As};
+use primitives::traits::{Zero, One, RefInto, As};
 use runtime_support::{StorageValue, StorageMap};
 
 pub mod voting;
@@ -568,7 +568,8 @@ mod tests {
 	pub use super::*;
 	pub use runtime_io::with_externalities;
 	pub use substrate_primitives::H256;
-	use primitives::{HasPublicAux, Identity, BuildExternalities};
+	use primitives::BuildExternalities;
+	use primitives::traits::{HasPublicAux, Identity};
 	use primitives::testing::{Digest, Header};
 
 	impl_outer_dispatch! {
